# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  contact:
    name: Stellar-IndigoPay Team
    url: https://github.com/Stellar-IndigoPay/Stellar-IndigoPay
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server  
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LimitQuery:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorQuery:
      name: cursor
      in: query
      schema:
        type: string
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    RateLimitExceeded:
      description: The request has been rate-limited. All endpoints carry rate-limit response headers (X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset) and return 429 with a Retry-After header when the limit is exceeded. See the API documentation for per-endpoint tiers.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests allowed (or bucket capacity for token-bucket endpoints).
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in the current window (or tokens left in the bucket).
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch seconds when the limit resets (or next token becomes available).
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests — Try again later.
              retryAfter:
                type: integer
                description: Epoch seconds after which a retry may succeed.
                example: 1742169602
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        category:
          type: string
        location:
          type: string
        walletAddress:
          type: string
        goalXLM:
          type: string
        raisedXLM:
          type: string
        donorCount:
          type: integer
        co2OffsetKg:
          type: integer
        status:
          type: string
          enum: [active, completed, paused, rejected]
        rejectionReason:
          type: string
          nullable: true
        verified:
          type: boolean
        onChainVerified:
          type: boolean
        tags:
          type: array
          items:
            type: string
        aiSummary:
          type: string
          nullable: true
        aiSummaryGeneratedAt:
          type: string
          format: date-time
          nullable: true
        aiSummaryModel:
          type: string
          nullable: true
        aiSummarySourceHash:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProjectRequest:
      type: object
      required: [name, description, location, category, wallet_address]
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 120
        description:
          type: string
          minLength: 10
          maxLength: 5000
        location:
          type: string
          minLength: 2
          maxLength: 200
        category:
          type: string
          enum:
            [
              Reforestation,
              Solar Energy,
              Ocean Conservation,
              Clean Water,
              Wildlife Protection,
              Carbon Capture,
              Wind Energy,
              Sustainable Agriculture,
              Other,
            ]
        wallet_address:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        goal_xlm:
          type: number
          default: 0
        tags:
          type: array
          items:
            type: string
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: listProjects
      description: Retrieve a paginated list of projects with optional filters for category, status, verification, and search.
      parameters:
        - $ref: "#/components/parameters/LimitQuery"
        - $ref: "#/components/parameters/CursorQuery"
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused]
        - name: verified
          in: query
          schema:
            type: string
            enum: ["true"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated project list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
                  next_cursor:
                    type: string
                    nullable: true
                  has_more:
                    type: boolean
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      description: Register a new climate project on the platform with name, description, location, category, and wallet address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Project"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...